# docs/swagger.yaml
basePath: /
definitions:
...
  resources.CreateResources:
    properties:
      body:
        type: string
      description:
        type: string
      facebook_link:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      instagram_link:
        type: string
      linkedin_link:
        type: string
      name:
        maxLength: 200
        type: string
      pinterest_link:
        type: string
      twitter_link:
        type: string
      type:
        items:
          type: string
        type: array
      youtube_link:
        type: string
    required:
    - name
    type: object
  resources.UpdateResources:
    properties:
      body:
        type: string
      description:
        type: string
      facebook_link:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      instagram_link:
        type: string
      linkedin_link:
        type: string
      name:
        type: string
      pinterest_link:
        type: string
      twitter_link:
        type: string
      type:
        items:
          type: string
        type: array
      youtube_link:
        type: string
    type: object
...
  user.UserUpdate:
    properties:
      audience_interests:
        type: string
      average_likes:
        type: number
      avg_comments:
        type: number
      avg_likes:
        type: number
      avg_reel_plays:
        type: number
      avg_views:
        type: number
      bio:
        type: string
      company_account:
        type: string
      email:
        type: string
      engagement_rate:
        type: number
      followers:
        type: number
      gender_split:
        type: string
      influencer_ig_name:
        type: string
      location:
        type: string
      manager_phone_number:
        type: string
      name:
        type: string
      password:
        default: "null"
        type: string
      phone_number:
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
      popular_posts:
        type: string
      provider:
        type: string
      role:
        type: string
      total_posts:
        type: number
    type: object
  utils.Base64Struct:
    properties:
      base64:
        type: string
    type: object
  utils.DefaultParam:
    properties:
      key:
        type: string
      value: {}
    type: object
  utils.FilterObj:
    properties:
      field_name:
        type: string
      field_type:
        type: string
      operation:
        type: string
      value:
        type: string
      values:
        items: {}
        type: array
    type: object
  utils.Pagination:
    properties:
      current_page_no:
        type: integer
      per_page:
        type: integer
      sort:
        type: string
      total_elements:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.RequestObj:
    properties:
      default_param:
        items:
          $ref: '#/definitions/utils.DefaultParam'
        type: array
      filter:
        items:
          $ref: '#/definitions/utils.FilterObj'
        type: array
      glob_operation:
        type: string
      page_no:
        type: integer
      per_page:
        type: integer
      sort:
        type: string
    type: object
  utils.ResponseObj:
    properties:
      data: {}
      error: {}
      pagination:
        $ref: '#/definitions/utils.Pagination'
      response_code:
        type: integer
      response_msg:
        type: string
    type: object
host: http://103.168.56.249:8080
info:
  contact: {}
  description: This is a sample API with Fiber and Swagger
  title: Ramen API
  version: "1.0"
paths:
...
  /resources:
    post:
      consumes:
      - application/json
      description: Create resources
      parameters:
      - description: CreateReference
        in: body
        name: resources
        required: true
        schema:
          $ref: '#/definitions/resources.CreateResources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Create resources
      tags:
      - Resources
  /resources/{id}:
    delete:
      consumes:
      - application/json
      description: Delete resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Delete resources
      tags:
      - Resources
    get:
      consumes:
      - application/json
      description: Get resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Get resources
      tags:
      - Resources
    put:
      consumes:
      - application/json
      description: Update resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateReference
        in: body
        name: resources
        required: true
        schema:
          $ref: '#/definitions/resources.UpdateResources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Update resources
      tags:
      - Resources
  /resources/list:
    post:
      consumes:
      - application/json
      description: List resources
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: List resources
      tags:
      - Resources
...
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      - description: UserUpdate
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - User
  /users/list:
    post:
      consumes:
      - application/json
      description: Get user list
      parameters:
      - description: RequestObj
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get user list
      tags:
      - User
  /users/me:
    get:
      consumes:
      - application/json
      description: Get user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - User
swagger: "2.0"